Fixed bug causing failed jobs when nmols doesn't change much

+++ Opt_ES/gemc_val/project.py
@@ -2514,7 +2514,7 @@ def plot_res_pymser(job, eq_col, results, name, box_name):
         n_ticks = 4
     elif n_ticks > 8:
         n_ticks = 8
-    step_size = int((eq_col.max()-eq_col.min()) / n_ticks)
+    step_size = (eq_col.max()-eq_col.min()) / n_ticks
     ax1.set_ylim(eq_col.min()-1, eq_col.max()+1)
     ax1.set_yticks(np.arange(int(eq_col.min()), int(eq_col.max() + step_size), step_size))
     # ax1.set_yticks(np.arange(0, eq_col.max() * 1.1, eq_col.max() / 10))
